planneroutdoor/requirements.txt
asgiref==3.8.1
Django==5.1.1
django-cors-headers==4.4.0
django-rest-auth==0.9.5
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
django-extensions==3.2.3
requests==2.32.3
inflection==0.5.1
mysqlclient==2.2.4
packaging==24.1
PyJWT==2.9.0
pytz==2024.2
PyYAML==6.0.2
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
uritemplate==4.1.1
pylint==3.3.1
